saki-plugin-sdk/pyproject.toml
[project]
name = "saki-plugin-sdk"
version = "2.0.0"
description = "Plugin SDK for Saki executor – base classes, IPC protocol, and built-in strategies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.2",
    "pydantic>=2.7.0",
    "pyzmq>=26.2.0",
    "pyyaml>=6.0",
    "saki-ir",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/saki_plugin_sdk"]

[tool.uv.sources]
saki-ir = { path = "../shared/saki-ir/python", editable = true }
